feature: Add summary support in AddSpentTimeOptions
This commit is adding a `summary` field in the `AddSpentTimeOptions`
structure.

This structure is used to call the add spent API endpoints for merge
requests and issues. Today we can only pass the duration of the time
spent when calling gitlab API, however the 2 endpoints also accept a
`summary` optional field.

Adding this new field in the structure will allow users to set a
specific `summary` value while calling gitlab API.

[1]: https://docs.gitlab.com/ee/api/merge_requests.html#add-spent-time-for-a-merge-request
[2]: https://docs.gitlab.com/ee/api/issues.html#add-spent-time-for-an-issue